Fix accuracy predictions.

+++ model/model_fn.py
@@ -83,7 +83,7 @@ def model_fn(mode, inputs, params, reuse=False):
     with tf.variable_scope('model', reuse=reuse):
         # Compute the output distribution of the model and the predictions
         logits = build_model(is_training, inputs, params)
-        predictions = tf.one_hot(tf.argmax(logits, 1), params.num_labels)
+        predictions = tf.cast(tf.one_hot(tf.argmax(logits, 1), params.num_labels), tf.int64)
 
     # Define loss
     loss = tf.losses.softmax_cross_entropy(onehot_labels=labels, logits=logits) + tf.losses.get_regularization_loss()
